pyproject.toml
[project]
name = "aiaxemind"
version = "0.1.0"
description = "Socratic AI Mentor for Programming Education"
readme = "README.md"
requires-python = "==3.12.*"
authors = [
    { name = "AiaxeMind Team" }
]

dependencies = [
    # Core Framework
    "fastapi==0.135.3",
    "uvicorn[standard]==0.43.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    # Database & Migrations
    "sqlalchemy==2.0.49",
    "alembic==1.18.4",
    "asyncpg==0.30.0",
    # LLM & AI Stack
    "langchain==1.2.15",
    "langchain-core==1.2.26",
    "langchain-openai==1.1.12",
    # Document Processing (torch installed separately, see README)
    "pymupdf==1.27.2.2",
    "python-docx==1.2.0",
    "sentence-transformers==5.3.0",
    # Vector Database
    "qdrant-client==1.17.1",
    # Utilities
    "python-multipart==0.0.22",
    "websockets==16.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.8",
    "pytest-cov==4.1.0",
    "httpx==0.27.2",
    "ruff==0.5.7",
    "mypy==1.11.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "F821"]  # F821: Forward references

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["src/models/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
